fixed CommonObjectDataSubRecord.field_2_objectId to be unsigned, also fixed HSSFCell.findCellComment to handle sheets with more than 65536 comments
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@782398 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/model/CommentShape.java
@@ -74,10 +74,10 @@ public class CommentShape extends TextboxShape {
      private NoteRecord createNoteRecord( HSSFComment shape, int shapeId )
     {
         NoteRecord note = new NoteRecord();
-        note.setColumn((short)shape.getColumn());
-        note.setRow((short)shape.getRow());
+        note.setColumn(shape.getColumn());
+        note.setRow(shape.getRow());
         note.setFlags(shape.isVisible() ? NoteRecord.NOTE_VISIBLE : NoteRecord.NOTE_HIDDEN);
-        note.setShapeId((short)shapeId);
+        note.setShapeId(shapeId);
         note.setAuthor(shape.getAuthor() == null ? "" : shape.getAuthor());
         return note;
     }
