+++ src/java/org/apache/poi/hssf/model/PictureShape.java
@@ -100,7 +100,7 @@ public class PictureShape
         CommonObjectDataSubRecord c = new CommonObjectDataSubRecord();
         c.setObjectType((short) ((HSSFSimpleShape)shape).getShapeType());
 //        c.setObjectId((short) ( 1 ));
-        c.setObjectId((short) ( shapeId ));
+        c.setObjectId(shapeId);
         c.setLocked(true);
         c.setPrintable(true);
         c.setAutofill(true);
